move migration logic to extensions/authentication/src/migration/
